agrego persistencia segundo parcial 2022

+++ src/main/java/models/segundoParcial2023/Deportista.java
@@ -8,7 +8,7 @@ import java.util.List;
 public class Deportista {
     @Id @GeneratedValue
     private Long id;
-    @Column(name="altura", columnDefinition="DECIMAL")
+    @Column(name="altura", columnDefinition="DOUBLE")
     private Double altura;
     @Column(name="apellido")
     private String apellido;
